% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_datasets.R
\name{combine_datasets}
\alias{combine_datasets}
\title{Gets Data from warren nolan}
\usage{
combine_datasets(kenpom_data, warrennolan_data)
}
\arguments{
\item{kenpom_data}{The data from kenpom}

\item{warrennolan_data}{the data from warren nolan}
}
\value{
combined data.
}
\description{
Gets Data from warren nolan
}
