% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_years.R
\docType{data}
\name{combined_years}
\alias{combined_years}
\title{3 Year SOR and SOR rank}
\format{
\describe{
\item{Team}{The name of the team}
\item{SOR}{The SOR for the given season}
\item{ranking}{The SOR ranking for the given year}
\item{year}{The year the other data is from}
}
}
\usage{
combined_years
}
\description{
The sor and sor ranking of all college basketball teams over 3 seasons
}
\examples{
data(combined_years)
head(combined_years)
}
\keyword{datasets}
