% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_and_format_kenpom_data.R
\name{fetch_and_format_kenpom_data}
\alias{fetch_and_format_kenpom_data}
\title{Gets Kenpom data and formats it in our usable form}
\usage{
fetch_and_format_kenpom_data(year)
}
\arguments{
\item{year}{the year to run obtain the kenmpom data}
}
\value{
Kenpom data of games played
}
\description{
Gets Kenpom data and formats it in our usable form
}
