% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_and_format_warrennolan_data.R
\name{fetch_and_format_wn_data}
\alias{fetch_and_format_wn_data}
\title{Gets Data from warren nolan}
\usage{
fetch_and_format_wn_data(year, team = Team)
}
\arguments{
\item{year}{the year to run obtain the Warren Nolan data}

\item{team}{the team that is being filtered}
}
\value{
Warren Nolan data
}
\description{
Gets Data from warren nolan
}
