% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_names.R
\name{match_names}
\alias{match_names}
\title{Matches names of teams properly}
\usage{
match_names(name, names_to_match)
}
\arguments{
\item{name}{one name}

\item{names_to_match}{the name to match}
}
\value{
The matched names
}
\description{
Matches names of teams properly
}
