% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_kenpom_row.R
\name{parse_kenpom_row}
\alias{parse_kenpom_row}
\title{Gets Kenpom data and formats it in our usable form}
\usage{
parse_kenpom_row(row)
}
\arguments{
\item{year}{the year to run obtain the kenmpom data}
}
\value{
Kenpom data of games played
}
\description{
Gets Kenpom data and formats it in our usable form
}
