% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankings_2023.R
\docType{data}
\name{rankings_2023}
\alias{rankings_2023}
\title{SOR rankings for 2023}
\format{
\describe{
\item{Team}{The name of the team}
\item{SOR}{The team's strenght of record}
\item{ranking}{The team's strenght of record ranking}
}
}
\usage{
rankings_2023
}
\description{
A strength of record ranking for D1 basketball teams in 2023
}
\keyword{datasets}
