% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sor_points.R
\name{sor_points}
\alias{sor_points}
\title{Gets the sor points for the home team}
\usage{
sor_points(score, opp_score, opp_net, number_of_teams, home, neutral = FALSE)
}
\arguments{
\item{score}{The team's own score}

\item{opp_score}{The opposing team's score}

\item{opp_net}{The opposing team's NET ranking}

\item{number_of_teams}{The number of teams in that college basketball season}

\item{home}{TRUE if team is home team, FALSE if team is away team}

\item{neutral}{TRUE if game is played in a neutral location, FALSE if not}
}
\value{
The strength of record points gained or lost by the home team
}
\description{
This returns the strength of record points for teams with the
following formula:
points gained in win = ((number of teams worse in NET than team you beat) +
(10 if neutral or 20 if away)) \ ((number of teams + 19)
}
\details{
This returns a 1 if a team beats the number 1 team on the road and returns
a 0 if the team beats the worst team at home, and a 1 if a team beats the
best team on the road

The number of point lost in a loss in simply 1 - the points to be gained with
a win. This means that losing to the best team on the road is a loss of 0,
and losing to the worst team at home is a loss of 1.
}
