% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_sor.R
\name{sum_sor}
\alias{sum_sor}
\title{Sums sor scores for each team.}
\usage{
sum_sor(
  home_team = combined_data$Home_team,
  away_team = combined_data$Away_team,
  home_sor = combined_data$home_sor,
  away_sor = combined_data$away_sor
)
}
\arguments{
\item{home_team}{the home team in a game}

\item{away_team}{the away team in a game}

\item{home_sor}{the home team's sor in a game}

\item{away_sor}{the away team's sor in a game}
}
\value{
a dataframe with every team and their sor
}
\description{
Sums sor scores for each team.
}
