% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_and_format_kenpom_data.R
\name{fetch_and_format_kenpom_data}
\alias{fetch_and_format_kenpom_data}
\title{Gets Kenpom data and formats it in our usable form}
\usage{
fetch_and_format_kenpom_data(
  year,
  away_team = Away_team,
  home_team = Home_team
)
}
\arguments{
\item{year}{the year to run obtain the kenmpom data}

\item{away_team}{The away team}

\item{home_team}{The home team}
}
\description{
Gets Kenpom data and formats it in our usable form
}
