% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_combined_data.R
\name{get_combined_data}
\alias{get_combined_data}
\title{Gets combined data to compute sor record}
\usage{
get_combined_data(year)
}
\arguments{
\item{year}{The year to get the data from}
}
\value{
combined data.
}
\description{
Gets combined data to compute sor record
}
