% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_sor.R
\name{sum_sor}
\alias{sum_sor}
\title{Sums sor for each team.}
\usage{
sum_sor(year, games_date)
}
\arguments{
\item{year}{The year to get the data from. Only works from 2021}

\item{games_date}{the cutoff date for the games played.}
}
\value{
a data frame with every team and their sor
}
\description{
Sums sor for each team.
}
